requirements.txt
flask==3.1.2
psycopg2-binary==2.9.10
redis==6.4.0
google-generativeai==0.8.5